requirements.txt
python-terrier==0.9.2
python-dotenv~=1.0.0
pandas~=2.0.3